seperated out message methods and refactored game.rb
